+++ src/containers/mod.rs
@@ -1,3 +1,4 @@
+pub mod area;
 pub mod column;
 pub mod frame;
 pub mod row;