+++ src/prelude.rs
@@ -6,6 +6,7 @@
 //! ```
 
 pub use crate::Apply;
+pub use crate::ColorExt;
 pub use crate::DesignSlots;
 pub use crate::Styled;
 pub use crate::StyledTheme;
@@ -13,6 +14,7 @@ pub use crate::ThemeExt;
 pub use crate::WebPalette;
 pub use crate::{rgb, rgba};
 
+pub use crate::StyledArea;
 pub use crate::StyledButton;
 pub use crate::StyledCheckbox;
 pub use crate::StyledColumn;